+++ src/RadioButtonCollection.php
@@ -109,6 +109,24 @@ class RadioButtonCollection extends AbstractSetFormValueElement implements Radio
     }
 
     /**
+     * Returns the selected radio button or null if no radio button is selected.
+     *
+     * @since 2.2.0
+     *
+     * @return RadioButtonInterface|null The selected radio button or null if no radio button is selected.
+     */
+    public function getSelectedRadioButton(): ?RadioButtonInterface
+    {
+        foreach ($this->radioButtons as $radioButton) {
+            if ($radioButton->getValue() === $this->value) {
+                return $radioButton;
+            }
+        }
+
+        return null;
+    }
+
+    /**
      * Returns the element value.
      *
      * @since 1.0.0