+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Unreleased
+- Added getSelectedRadioButton method to RadioButtonCollection.
+- Added getSelectedOption method to Select.
 
 ## [2.1.0] - 2019-06-14
 - Made RadioButtonCollection and Select countable.